Bucket retention lock integration tests
Replace the single large integration test with individual tests
equivalent to the Java tests.

It's expected that some buckets may not be able to be deleted at the
end of a test now, particularly if tests fail. The clean-up program
is therefore updated to try to remove event-based holds on both
objects and buckets if it runs into a problem - the next clean-up
run should then be able to delete the bucket, after any retention
period has passed.